Move AppSettings and related enums to a config namespace
